refactor: fix docstring

+++ src/pymmcore_widgets/_mda/_zstack_widget.py
@@ -296,7 +296,7 @@ class ZStackWidget(QGroupBox):
         """Return the current number of images in the z-stack."""
         step = self._zstep_spinbox.value()
         _range = cast("ZPicker", self._zmode_tabs.currentWidget()).z_range()
-        return round((_range / step) + 1)
+        return int(round((_range / step) + 1))
 
     def set_state(self, z_plan: dict) -> None:
         """Set the state of the widget from a dictionary."""
